refactor(app): extract line/search scan orchestrators (phase 5c)
Move start_incremental_line_scan, process_line_scan, process_line_scan_batch,
finish_line_scan_ok, finish_line_scan_with_error, open_goto_line_if_active,
process_search_scan, process_search_scan_batch, finish_search_scan out
of buffer_management.rs into a new app/scan_orchestrators.rs sibling.
~265 lines.

These drive the chunked-scan subsystems (LineScan, SearchScan from
phase 2) one batch per render frame, coordinating with the buffer
registry, tokio runtime, and status message.

buffer_management.rs goes from 2196 -> 1929 lines (-267).
2026 tests still pass.
