+++ crates/fresh-editor/src/app/mod.rs
@@ -47,6 +47,7 @@ mod prompt_actions;
 mod recovery_actions;
 mod regex_replace;
 mod render;
+mod scan_orchestrators;
 mod scroll_sync;
 mod scrollbar_input;
 mod scrollbar_math;
